+++ FirebasePerformance/Sources/AppActivity/FPRAppActivityTracker.m
@@ -34,9 +34,6 @@ static FPRMemoryGaugeData *gAppStartMemoryGaugeData = nil;
 static BOOL isActivePrewarm = NO;
 
 NSString *const kFPRAppStartTraceName = @"_as";
-NSString *const kFPRAppStartStageNameTimeToUI = @"_astui";
-NSString *const kFPRAppStartStageNameTimeToFirstDraw = @"_astfd";
-NSString *const kFPRAppStartStageNameTimeToUserInteraction = @"_asti";
 NSString *const kFPRAppTraceNameForegroundSession = @"_fs";
 NSString *const kFPRAppTraceNameBackgroundSession = @"_bs";
 NSString *const kFPRAppCounterNameTraceEventsRateLimited = @"_fstec";
@@ -210,10 +207,6 @@ NSString *const kFPRAppCounterNameActivePrewarm = @"_fsapc";
   dispatch_once(&onceToken, ^{
     self.appStartTrace = [[FIRTrace alloc] initInternalTraceWithName:kFPRAppStartTraceName];
     [self.appStartTrace startWithStartTime:appStartTime];
-    [self.appStartTrace startStageNamed:kFPRAppStartStageNameTimeToUI startTime:appStartTime];
-
-    // Start measuring time to first draw on the App start trace.
-    [self.appStartTrace startStageNamed:kFPRAppStartStageNameTimeToFirstDraw];
   });
 
   // If ever the app start trace had it life in background stage, do not send the trace.
@@ -234,7 +227,6 @@ NSString *const kFPRAppCounterNameActivePrewarm = @"_fsapc";
   // Start measuring time to make the app interactive on the App start trace.
   static BOOL TTIStageStarted = NO;
   if (!TTIStageStarted) {
-    [self.appStartTrace startStageNamed:kFPRAppStartStageNameTimeToUserInteraction];
     TTIStageStarted = YES;
 
     // Assumption here is that - the app becomes interactive in the next runloop cycle.